fix: parse through CSV to add attributes to database
