+++ db/database.py
@@ -1,4 +1,6 @@
 import sqlite3
+import csv
+import sys
 
 connection = sqlite3.connect("./db/pineapplestore.db")
 cursor = connection.cursor()
@@ -22,8 +24,8 @@ cursor.execute(create_history_table)
 
 create_inventory_table = '{}{}{}{}'.format(
     'CREATE TABLE IF NOT EXISTS',
-    ' inventory(id INTEGER PRIMARY KEY, sku text, upc INTEGER',
-    ' rando text, product text, description text, price FLOAT', 
+    ' inventory(id INTEGER PRIMARY KEY, sku text, upc INTEGER,',
+    ' rando text, product text, description text, price FLOAT,', 
     ' size text, color text, amt INTEGER);'
 )
 cursor.execute(create_inventory_table)
@@ -35,10 +37,12 @@ cursor.execute('INSERT OR REPLACE INTO user VALUES(4, "kangning_chen", "asdfg", 
 cursor.execute('INSERT OR REPLACE INTO user VALUES(5, "yunqi_qian", "qwerty", "Ann Arbor", "Male", "12", "30","30","XL");')
 cursor.execute('INSERT OR REPLACE INTO user VALUES(6, "tayloir_thompson", "aqwerva", "Ann Arbor", "Male", "12", "30","30","XL");')
 
-# cursor.execute('INSERT OR REPLACE INTO inventory VALUES(1, "tshirt", 10999.99);')
-# cursor.execute('INSERT OR REPLACE INTO inventory VALUES(2, "shorts", 45.99);')
-# cursor.execute('INSERT OR REPLACE INTO inventory VALUES(3, "cardigan", 599.99);')
-# cursor.execute('INSERT OR REPLACE INTO inventory VALUES(4, "boots", 20050.99);')
+with open("./db/pineapple_inventory.csv", "rt") as f:
+    rows = csv.reader(f)
+    next(rows) # Skip the header row.
+    for row in rows:
+        query = "INSERT OR REPLACE INTO inventory VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
+        cursor.execute(query, row)
 
 cursor.execute('INSERT OR REPLACE INTO purchase_history VALUES(1, "tshirt", 1, 1);')
